mcp-server/requirements.txt
mcp>=1.0.0
PyGithub>=1.59.0
python-dotenv>=1.0.0
